Choosing the right url with locale when open privacy policy or terms
Privacy policy and terms and conditions links are published as github pages and have the following template: `https://myhoroskope.app/privacy_policy/:lang.html`.
For instance, for English, it would be `https://myhoroskope.app/privacy_policy/en.html`

So, in order to choose the right policy, we need to provide the right locate identifier.