Appodeal integration and user ads consent
In order to be consistent with GDPR and other data privacy protection laws,
we need to ask user for a permission to use their ad tracking id.

Some of the ad networks implemented such dialogs.
We are using one from Appodeal.

We ask the permission namely when the user has clicked over four cards and ready to see the ads.
Then we ash the user to give consent and show the "Give consent" button.

This commit also contains Appodeal SDK integration.

What has been added:
* actualized `app-ads.txt`
* added localizations for the user consent
* android SDK version updated to 21 (requirement of Appodeal)
* removed `google_admob` and its id from the manifest